Move public headers into a separate include/woff2 directory. #49
