+++ src/woff2_compress.cc
@@ -17,7 +17,7 @@
 #include <string>
 
 #include "file.h"
-#include "./woff2_enc.h"
+#include <woff2/encode.h>
 
 
 int main(int argc, char **argv) {
